moving PersonalizeProvider type into separate file
- Moving PersonalizeProvider type into separate file
- Import it from the main Analytics.ts type file
