+++ packages/analytics/src/types/Providers/AmazonPersonalizeProvider.ts
@@ -1,0 +1,7 @@
+export interface PersonalizeAnalyticsEvent {
+	eventType?: 'Identify' | 'MediaAutoTrack';
+	userId?: string;
+	properties?: {
+		[key: string]: string;
+	};
+}
